requirementes.txt
Django==1.6.5
MySQL-python==1.2.4
PIL==1.1.7
South==1.0
app-version==0.2.1
django-appconf==0.6
django-braces==1.4.0
django-guardian==1.2.4
django-permission==0.8.2
django-tastypie==0.11.1
psycopg2==2.5.3
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
tolerance==0.1.2
wsgiref==0.1.2
